Accumulated

Calculates DateTime when reach accumulated total.

dts_accumulated(
  x,
  dtt = "DateTime",
  colname = "Value",
  accum = c(Hatch = 320),
  units = "days"
)

Arguments

x
A dts data frame

dtt
A string specifying the column with the Date or POSIXct values.

colname
A character vector specifying the column(s) with the other values.

accum
A uniquely named vector of the accumulation thresholds.

units
A string of the accumulation units. If the accumulation units differ from the dtt units the accumulation units are automatically converted.

Value

The modified DateTime series data frame.

Examples

dts_accumulated(dts_data[1:5,], colname = "Value2", 
  accum = c(Hatch = 40), units = "hours")
#> # A tibble: 5 x 5
#>   DateTime            Value Value2 Value3 Hatch              
#>   <dttm>              <dbl>  <dbl>  <dbl> <dttm>             
#> 1 1999-11-29 22:00:00  -9     20.1   4.10 1999-11-30 00:00:00
#> 2 1999-11-29 23:00:00  NA     20.3   4.07 1999-11-30 01:00:00
#> 3 1999-11-30 00:00:00  NA     19.9   4.10 1999-11-30 02:00:00
#> 4 1999-11-30 01:00:00  NA     20.5   4.11 NA                 
#> 5 1999-11-30 02:00:00  10.3   20.4   4.12 NA