Completes a data time series

Completes a data time series

dts_complete(x, dtt = "DateTime", colname = dts_colnames(x),
  from = min(x[[dtt]]), to = max(x[[dtt]]), floor = TRUE,
  unique = TRUE, sort = TRUE, units = dttr::dtt_units(x[[dtt]]),
  .dts_fun = mean, ...)

Arguments

x
A Date or POSIXct vector.

dtt
A string specifying the column with the Date or POSIXct values.

colname
A character vector specifying the column(s) with the other values.

from
A Date or POSIXct scalar of the start date.

to
A Date or POSIXct scalar of the to date.

floor
A flag indicating whether to floor the values.

unique
A flag indicating whether to return distinct values.

sort
A flag indicating whether the returned vector must be sorted.

units
A string of the units.

.dts_fun
A function which returns a scalar.

...
Unused.

Value

The completed data time series.

Examples

dts_complete(dts_data[c(1,3),])

#> # A tibble: 3 x 4
#>   DateTime            Value Value2 Value3
#>   <dttm>              <dbl>  <dbl>  <dbl>
#> 1 1999-11-29 22:00:00    -9   20.1   4.10
#> 2 1999-11-29 23:00:00    NA   NA    NA   
#> 3 1999-11-30 00:00:00    NA   19.9   4.10