Distribution

Calculates distribution by DateTime.

dts_distribution(
  x,
  dtt = "DateTime",
  colname = "Distribution",
  units = "days",
  normalize = TRUE,
  .dts_fun = "dnorm",
  .timing,
  ...
)

Arguments

x
A dts data frame

dtt
A string specifying the column with the Date or POSIXct values.

colname
A character vector specifying the column(s) with the other values.

units
A string of the time units for the variation (3rd and more) arguments of .dts_fun.

normalize
A flag indicating whether to normalize the values (ensure they sum to 1) or the proportion of the summed values to retain before renormalizing.

.dts_fun
A string of the name of a function which takes a numeric vector as its first argument (the values to calculate the distribution for), a numeric scalar as it second argument (the timing of the event) and one or more numeric scalars (representing the variation in the timing of the event).

.timing
A Date or POSIXct vector each element of which will be converted into a numeric value to be passed as the second argument to .dts_fun.

...
Additional arguments passed as named arguments to .dts_fun which should all be numeric vectors of the same length as .dtt or numeric scalars (which are reused).

Value

The modified DateTime series data frame with a column of the distribution.

Details

The date time values in x are converted into a numeric vector and passed as the first argument to .dts_fun.

See also

Examples

dts_distribution(dts_data[1:10,], .timing = dts_data$DateTime[2], sd = 2, units = "hours")
#> # A tibble: 10 × 5
#>    DateTime            Value Value2 Value3 Distribution
#>    <dttm>              <dbl>  <dbl>  <dbl>        <dbl>
#>  1 1999-11-29 22:00:00  -9     20.1   4.10    0.227    
#>  2 1999-11-29 23:00:00  NA     20.3   4.07    0.257    
#>  3 1999-11-30 00:00:00  NA     19.9   4.10    0.227    
#>  4 1999-11-30 01:00:00  NA     20.5   4.11    0.156    
#>  5 1999-11-30 02:00:00  10.3   20.4   4.12    0.0835   
#>  6 1999-11-30 03:00:00  10.3   NA    NA       0.0348   
#>  7 1999-11-30 04:00:00  10.3   20.3   4.11    0.0113   
#>  8 1999-11-30 05:00:00  10.2   20.6   4.10    0.00286  
#>  9 1999-11-30 06:00:00  10.3   20.5   4.09    0.000562 
#> 10 1999-11-30 07:00:00  10.2   NA     4.10    0.0000863