Floors the Date or POSIXct column of a date time series data frame.

Floors the Date or POSIXct column of a date time series data frame.

dts_floor(x, dtt = "DateTime", units = dttr::dtt_units(x[[dtt]]))

Arguments

x
A Date or POSIXct vector.

dtt
A string specifying the column with the Date or POSIXct values.

units
A string of the units.

Value

The completed data time series.

Examples

dts_floor(dts_data[c(1,3),], units = "years")

#> # A tibble: 2 x 4
#>   DateTime            Value Value2 Value3
#>   <dttm>              <dbl>  <dbl>  <dbl>
#> 1 1999-01-01 00:00:00    -9   20.1   4.10
#> 2 1999-01-01 00:00:00    NA   19.9   4.10